requirements.txt
python-telegram-bot
PyMuPDF
httpx
python-dotenv